requirements.txt
numpy==1.19.5
pandas==1.1.5
scipy==1.5.4
matplotlib==3.3.4
seaborn==0.11.1
scikit-learn==0.24.1
tensorflow-gpu==2.4.1
joblib==1.0.1
